fix: moving groupId and package from software.aws.lambda to software.amazon.lambda
