+++ docs/content/index.mdx
@@ -14,7 +14,7 @@ Powertools is available in Maven Central. You can use your favourite dependency 
 <dependencies>
     ...
     <dependency>
-        <groupId>software.aws.lambda</groupId>
+        <groupId>software.amazon.lambda</groupId>
         <artifactId>aws-lambda-powertools-java</artifactId>
         <version>YOUR_REQUIRED_VERSION</version>
     </dependency>
@@ -28,7 +28,7 @@ repositories {
 }
 
 dependencies {
-    powertools 'software.aws.lambda:aws-lambda-powertools-java:YOUR_REQUIRED_VERSION'
+    powertools 'software.amazon.lambda:aws-lambda-powertools-java:YOUR_REQUIRED_VERSION'
 }
 ```
