+++ src/main/java/software/amazon/lambda/logging/PowerToolsLogging.java
@@ -1,0 +1,13 @@
+package software.amazon.lambda.logging;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface PowerToolsLogging {
+
+    boolean logEvent() default false;
+}
