--- src/main/java/software/aws/lambda/logging/PowerLogger.java
@@ -1,10 +1,0 @@
-package software.aws.lambda.logging;
-
-import org.apache.logging.log4j.ThreadContext;
-
-public class PowerLogger {
-
-    public static void customKey(String key, String value) {
-        ThreadContext.put(key, value);
-    }
-}
