--- src/main/java/software/aws/lambda/logging/PowerToolsLogging.java
@@ -1,13 +1,0 @@
-package software.aws.lambda.logging;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface PowerToolsLogging {
-
-    boolean logEvent() default false;
-}
