--- src/main/java/software/aws/lambda/tracing/PowerToolTracing.java
@@ -1,14 +1,0 @@
-package software.aws.lambda.tracing;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface PowerToolTracing {
-    String namespace() default "";
-    boolean captureResponse() default true;
-    boolean captureError() default true;
-}
