+++ src/test/java/software/amazon/lambda/handlers/PowerToolDisabled.java
@@ -1,0 +1,12 @@
+package software.amazon.lambda.handlers;
+
+import com.amazonaws.services.lambda.runtime.Context;
+import com.amazonaws.services.lambda.runtime.RequestHandler;
+
+public class PowerToolDisabled implements RequestHandler<Object, Object> {
+
+    @Override
+    public Object handleRequest(Object input, Context context) {
+        return null;
+    }
+}
