+++ src/test/java/software/amazon/lambda/handlers/PowerToolDisabledForStream.java
@@ -1,0 +1,15 @@
+package software.amazon.lambda.handlers;
+
+import java.io.InputStream;
+import java.io.OutputStream;
+
+import com.amazonaws.services.lambda.runtime.Context;
+import com.amazonaws.services.lambda.runtime.RequestStreamHandler;
+
+public class PowerToolDisabledForStream implements RequestStreamHandler {
+
+    @Override
+    public void handleRequest(InputStream input, OutputStream output, Context context) {
+
+    }
+}
