+++ src/test/java/software/amazon/lambda/handlers/PowerToolLogEventEnabled.java
@@ -1,0 +1,14 @@
+package software.amazon.lambda.handlers;
+
+import com.amazonaws.services.lambda.runtime.Context;
+import com.amazonaws.services.lambda.runtime.RequestHandler;
+import software.amazon.lambda.logging.PowerToolsLogging;
+
+public class PowerToolLogEventEnabled implements RequestHandler<Object, Object> {
+
+    @PowerToolsLogging(logEvent = true)
+    @Override
+    public Object handleRequest(Object input, Context context) {
+        return null;
+    }
+}
